Fix issue of only 1/4 particles being updated
The issue was that the indices were being scaled by their global
work-id * sizeof(float) * 3. I  didn't really think about this because 3
* sizeof(float) is the size of the glm::vec3s in memory. The problem is
OpenCL is already working with floats so it does that pointer arithmetic
and doesn't need the sizeof(float) scale.

This fixes a lot of crashes (bad access on GPU), and now it works
correctly